# contracts/openapi.yaml
components:
  schemas:
    Body_import_map_api_v1_maps_import_post:
      properties:
        file:
          format: binary
          title: File
          type: string
      required:
      - file
      title: Body_import_map_api_v1_maps_import_post
      type: object
    ComponentStatus:
      enum:
      - HEALTHY
      - DEGRADED
      - UNAVAILABLE
      - DISABLED
      title: ComponentStatus
      type: string
    ExperimentCreateRequest:
      additionalProperties: false
      properties:
        map_id:
          anyOf:
          - minLength: 1
            type: string
          - type: 'null'
          title: Map Id
        scenario_id:
          format: uuid
          title: Scenario Id
          type: string
        workspace_id:
          format: uuid
          title: Workspace Id
          type: string
      required:
      - workspace_id
      - scenario_id
      title: ExperimentCreateRequest
      type: object
    ExperimentStatus:
      enum:
      - CREATED
      - PREPARING
      - READY
      - RUNNING
      - PAUSED
      - STOPPING
      - COMPLETED
      - FAILED
      title: ExperimentStatus
      type: string
    ExperimentView:
      additionalProperties: false
      properties:
        experiment_id:
          format: uuid
          title: Experiment Id
          type: string
        simulation_time_ms:
          minimum: 0.0
          title: Simulation Time Ms
          type: integer
        speed_multiplier:
          exclusiveMinimum: 0.0
          title: Speed Multiplier
          type: number
        status:
          $ref: '#/components/schemas/ExperimentStatus'
        workspace_id:
          format: uuid
          title: Workspace Id
          type: string
      required:
      - experiment_id
      - workspace_id
      - status
      - simulation_time_ms
      - speed_multiplier
      title: ExperimentView
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    HealthResponse:
      additionalProperties: false
      properties:
        service:
          const: trafficverse-api
          default: trafficverse-api
          title: Service
          type: string
        status:
          const: ok
          default: ok
          title: Status
          type: string
      title: HealthResponse
      type: object
    MapImportJob:
      additionalProperties: false
      properties:
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
        errors:
          default: []
          items:
            type: string
          title: Errors
          type: array
        job_id:
          format: uuid
          title: Job Id
          type: string
        map_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Map Id
        status:
          enum:
          - PENDING
          - RUNNING
          - SUCCEEDED
          - FAILED
          title: Status
          type: string
      required:
      - job_id
      - status
      title: MapImportJob
      type: object
    MapSummary:
      additionalProperties: false
      properties:
        carla_map:
          anyOf:
          - minLength: 1
            type: string
          - type: 'null'
          title: Carla Map
        carla_version:
          anyOf:
          - minLength: 1
            type: string
          - type: 'null'
          title: Carla Version
        display_name:
          anyOf:
          - minLength: 1
            type: string
          - type: 'null'
          title: Display Name
        files:
          default: []
          items:
            type: string
          title: Files
          type: array
        kind:
          default: core_run
          enum:
          - core_run
          - sumo
          title: Kind
          type: string
        manifest_available:
          default: true
          title: Manifest Available
          type: boolean
        map_id:
          minLength: 1
          title: Map Id
          type: string
        network_schema_version:
          minLength: 1
          title: Network Schema Version
          type: string
        sumo_begin_time_ms:
          default: 0
          minimum: 0.0
          title: Sumo Begin Time Ms
          type: integer
        sumo_config_file:
          anyOf:
          - minLength: 1
            type: string
          - type: 'null'
          title: Sumo Config File
        sumo_end_time_ms:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
          title: Sumo End Time Ms
        sumo_step_ms:
          anyOf:
          - exclusiveMinimum: 0.0
            type: integer
          - type: 'null'
          title: Sumo Step Ms
        validated:
          title: Validated
          type: boolean
        validation_errors:
          default: []
          items:
            type: string
          title: Validation Errors
          type: array
      required:
      - map_id
      - validated
      - network_schema_version
      title: MapSummary
      type: object
    ReadinessComponent:
      additionalProperties: false
      properties:
        component:
          minLength: 1
          title: Component
          type: string
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
        required:
          title: Required
          type: boolean
        status:
          $ref: '#/components/schemas/ComponentStatus'
      required:
      - component
      - status
      - required
      title: ReadinessComponent
      type: object
    ReadinessResponse:
      additionalProperties: false
      properties:
        components:
          items:
            $ref: '#/components/schemas/ReadinessComponent'
          title: Components
          type: array
        ready:
          title: Ready
          type: boolean
      required:
      - ready
      - components
      title: ReadinessResponse
      type: object
    SetSpeedRequest:
      additionalProperties: false
      properties:
        multiplier:
          exclusiveMinimum: 0.0
          maximum: 16.0
          title: Multiplier
          type: number
      required:
      - multiplier
      title: SetSpeedRequest
      type: object
    StopExperimentRequest:
      additionalProperties: false
      properties:
        reason:
          default: USER_REQUEST
          maxLength: 100
          minLength: 1
          title: Reason
          type: string
      title: StopExperimentRequest
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    WorkspaceActivitySample:
      additionalProperties: false
      properties:
        day:
          format: date
          title: Day
          type: string
        simulations:
          minimum: 0.0
          title: Simulations
          type: integer
      required:
      - day
      - simulations
      title: WorkspaceActivitySample
      type: object
    WorkspaceAutomationCount:
      additionalProperties: false
      properties:
        count:
          minimum: 0.0
          title: Count
          type: integer
        level:
          maxLength: 40
          minLength: 1
          title: Level
          type: string
      required:
      - level
      - count
      title: WorkspaceAutomationCount
      type: object
    WorkspaceCreateRequest:
      additionalProperties: false
      properties:
        description:
          default: ''
          maxLength: 1000
          title: Description
          type: string
        name:
          maxLength: 200
          minLength: 1
          title: Name
          type: string
      required:
      - name
      title: WorkspaceCreateRequest
      type: object
    WorkspaceOverview:
      additionalProperties: false
      description: Mock-backed workspace overview kept stable for a future real data
        source.
      properties:
        activity:
          items:
            $ref: '#/components/schemas/WorkspaceActivitySample'
          title: Activity
          type: array
        agent_count:
          minimum: 0.0
          title: Agent Count
          type: integer
        automation_counts:
          items:
            $ref: '#/components/schemas/WorkspaceAutomationCount'
          title: Automation Counts
          type: array
        failed_simulations:
          minimum: 0.0
          title: Failed Simulations
          type: integer
        map_count:
          minimum: 0.0
          title: Map Count
          type: integer
        preview_region:
          maxLength: 200
          minLength: 1
          title: Preview Region
          type: string
        recent_simulations:
          items:
            $ref: '#/components/schemas/WorkspaceRecentSimulation'
          title: Recent Simulations
          type: array
        runtime_hours:
          minimum: 0.0
          title: Runtime Hours
          type: number
        scenario_count:
          minimum: 0.0
          title: Scenario Count
          type: integer
        simulation_count:
          minimum: 0.0
          title: Simulation Count
          type: integer
        succeeded_simulations:
          minimum: 0.0
          title: Succeeded Simulations
          type: integer
        workspace_id:
          format: uuid
          title: Workspace Id
          type: string
      required:
      - workspace_id
      - map_count
      - agent_count
      - scenario_count
      - simulation_count
      - automation_counts
      - succeeded_simulations
      - failed_simulations
      - runtime_hours
      - activity
      - recent_simulations
      - preview_region
      title: WorkspaceOverview
      type: object
    WorkspaceRecentSimulation:
      additionalProperties: false
      properties:
        automation_summary:
          maxLength: 100
          minLength: 1
          title: Automation Summary
          type: string
        duration_ms:
          minimum: 0.0
          title: Duration Ms
          type: integer
        name:
          maxLength: 200
          minLength: 1
          title: Name
          type: string
        occurred_at:
          format: date-time
          title: Occurred At
          type: string
        status:
          enum:
          - SUCCEEDED
          - WARNING
          - FAILED
          title: Status
          type: string
      required:
      - name
      - status
      - occurred_at
      - duration_ms
      - automation_summary
      title: WorkspaceRecentSimulation
      type: object
    WorkspaceUpdateRequest:
      additionalProperties: false
      properties:
        description:
          default: ''
          maxLength: 1000
          title: Description
          type: string
        name:
          maxLength: 200
          minLength: 1
          title: Name
          type: string
      required:
      - name
      title: WorkspaceUpdateRequest
      type: object
    WorkspaceView:
      additionalProperties: false
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        description:
          default: ''
          maxLength: 1000
          title: Description
          type: string
        name:
          maxLength: 200
          minLength: 1
          title: Name
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
        workspace_id:
          format: uuid
          title: Workspace Id
          type: string
      required:
      - workspace_id
      - name
      - created_at
      - updated_at
      title: WorkspaceView
      type: object
info:
  title: TrafficVerse API
  version: 1.0.0
openapi: 3.1.0
paths:
  /api/v1/experiments:
    post:
      operationId: create_experiment_api_v1_experiments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExperimentCreateRequest'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Experiment
  /api/v1/experiments/{experiment_id}:
    get:
      operationId: get_experiment_api_v1_experiments__experiment_id__get
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          format: uuid
          title: Experiment Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Experiment
  /api/v1/experiments/{experiment_id}/pause:
    post:
      operationId: pause_experiment_api_v1_experiments__experiment_id__pause_post
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          format: uuid
          title: Experiment Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Pause Experiment
  /api/v1/experiments/{experiment_id}/prepare:
    post:
      operationId: prepare_experiment_api_v1_experiments__experiment_id__prepare_post
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          format: uuid
          title: Experiment Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Prepare Experiment
  /api/v1/experiments/{experiment_id}/resume:
    post:
      operationId: resume_experiment_api_v1_experiments__experiment_id__resume_post
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          format: uuid
          title: Experiment Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Resume Experiment
  /api/v1/experiments/{experiment_id}/speed:
    post:
      operationId: set_speed_api_v1_experiments__experiment_id__speed_post
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          format: uuid
          title: Experiment Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetSpeedRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Set Speed
  /api/v1/experiments/{experiment_id}/start:
    post:
      operationId: start_experiment_api_v1_experiments__experiment_id__start_post
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          format: uuid
          title: Experiment Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Start Experiment
  /api/v1/experiments/{experiment_id}/stop:
    post:
      operationId: stop_experiment_api_v1_experiments__experiment_id__stop_post
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          format: uuid
          title: Experiment Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StopExperimentRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Stop Experiment
  /api/v1/health:
    get:
      operationId: health_api_v1_health_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
          description: Successful Response
      summary: Health
  /api/v1/maps:
    get:
      operationId: maps_api_v1_maps_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MapSummary'
                title: Response Maps Api V1 Maps Get
                type: array
          description: Successful Response
      summary: Maps
  /api/v1/maps/import:
    post:
      operationId: import_map_api_v1_maps_import_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_map_api_v1_maps_import_post'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapImportJob'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Import Map
  /api/v1/maps/import/{job_id}:
    get:
      operationId: import_status_api_v1_maps_import__job_id__get
      parameters:
      - in: path
        name: job_id
        required: true
        schema:
          format: uuid
          title: Job Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapImportJob'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Import Status
  /api/v1/maps/{map_id}/manifest:
    get:
      operationId: map_manifest_api_v1_maps__map_id__manifest_get
      parameters:
      - in: path
        name: map_id
        required: true
        schema:
          title: Map Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Map Manifest Api V1 Maps  Map Id  Manifest Get
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Map Manifest
  /api/v1/maps/{map_id}/network:
    get:
      operationId: map_network_api_v1_maps__map_id__network_get
      parameters:
      - in: path
        name: map_id
        required: true
        schema:
          title: Map Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Map Network
  /api/v1/ready:
    get:
      operationId: ready_api_v1_ready_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadinessResponse'
          description: Successful Response
      summary: Ready
  /api/v1/workspaces:
    get:
      operationId: list_workspaces_api_v1_workspaces_get
      parameters:
      - in: query
        name: query
        required: false
        schema:
          anyOf:
          - maxLength: 200
            type: string
          - type: 'null'
          title: Query
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/WorkspaceView'
                title: Response List Workspaces Api V1 Workspaces Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Workspaces
    post:
      operationId: create_workspace_api_v1_workspaces_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceCreateRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Workspace
  /api/v1/workspaces/{workspace_id}:
    delete:
      operationId: delete_workspace_api_v1_workspaces__workspace_id__delete
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          format: uuid
          title: Workspace Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Workspace
    patch:
      operationId: update_workspace_api_v1_workspaces__workspace_id__patch
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          format: uuid
          title: Workspace Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceUpdateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Workspace
  /api/v1/workspaces/{workspace_id}/overview:
    get:
      operationId: workspace_overview_api_v1_workspaces__workspace_id__overview_get
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          format: uuid
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceOverview'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Workspace Overview